+++ react/universities/src/components/About.js
@@ -1,0 +1,12 @@
+const About = () => {
+  return (
+    <div className="container">
+      <div className="col">
+        <h1>A propos de l'application Universities</h1>
+        <p>Lorem ipsum dolor sit amet consectetur adipisicing elit. Repellat quisquam, recusandae repudiandae nobis, quis esse consequatur ipsam quod autem veniam quaerat facere deleniti quidem cum obcaecati, nemo commodi quos quas!</p>
+      </div>
+    </div>
+  );
+}
+
+export default About;
