+++ react/universities/src/services/UniversityData.js
@@ -1,5 +1,5 @@
 export default class UniversityData {
-  static base_url = 'https://coopernet.fr/totitied';
+  static base_url = 'http://universities.hipolabs.com';
   
   static getUniversities(country) {
     return fetch(`${this.base_url}/search?country=${country}`)
